Reorganize api file structure, scaffold MySqlService+Data route
