--- api-starter/src/api/services/test.ts
@@ -1,9 +1,0 @@
-export class TestService {
-
-    testGet(param : string) {
-
-        return `Test request with param ${param}`;
-
-    }
-
-}
