+++ api-starter/src/index.ts
@@ -2,8 +2,7 @@ import express from "express";
 import "module-alias/register";
 import { initialize } from "express-openapi";
 import { MySqlService } from "@services/MySqlService";
-import { TestService } from "@services/test";
-import apiDoc from "./api/apiDoc";
+import apiDoc from "./apiDoc";
 import swaggerUi from "swagger-ui-express";
 import Soda from "types/Soda";
 import bodyParser from "body-parser";
@@ -20,8 +19,6 @@ const app = express();
 
 const router = express.Router();
 
-const testService = new TestService();
-
 // Create a data consumer for the baltimore city dataset 
 // A Soda service will probably be created to facilitate all the soda stuff
 
@@ -34,32 +31,27 @@ consumer.query().limit(1000000000).withDataset("wsfq-mvij").where(soda.expr.eq("
     
 });
 
-router.get("/base", (req, res, next) => {
-    console.log("Router get called");
-    next();
-});
-
 app.use(router);
 app.use(bodyParser.json());
 
 const openapi = initialize({
     apiDoc: apiDoc,
     app,
-    paths: "./dist/api/routes",
+    paths: "./dist/routes",
     routesGlob: "**/*.{ts,js}",
     routesIndexFileRegExp: /(?:index)?\.[tj]s$/,
     dependencies: {
-        "testService": testService
+        "mySqlService": mySqlService
     }
 });
 
 
 // Setup swagger UI at /api using the doc on the openapi object
 
-app.use("/api",swaggerUi.serve,swaggerUi.setup(openapi.apiDoc));
+app.use("/doc",swaggerUi.serve,swaggerUi.setup(openapi.apiDoc));
 
-app.use(((err, req, res) => {
-    res.status(err.status).json(err);
-}) as express.ErrorRequestHandler);
+// app.use(((err, req, res) => {
+//     res.status(err.status).json(err);
+// }) as express.ErrorRequestHandler);
 
 app.listen(3000, () => console.log("Server listening on port 3000"));
